Conflicts:
	src/java/org/apache/cassandra/auth/Auth.java
	src/java/org/apache/cassandra/auth/PasswordAuthenticator.java
	src/java/org/apache/cassandra/config/CFMetaData.java
	src/java/org/apache/cassandra/config/Config.java
	src/java/org/apache/cassandra/config/DatabaseDescriptor.java
	src/java/org/apache/cassandra/config/Schema.java
	src/java/org/apache/cassandra/cql3/CQL3Type.java
	src/java/org/apache/cassandra/cql3/ColumnIdentifier.java
	src/java/org/apache/cassandra/cql3/Constants.java
	src/java/org/apache/cassandra/cql3/QueryProcessor.java
	src/java/org/apache/cassandra/cql3/Relation.java
	src/java/org/apache/cassandra/cql3/Sets.java
	src/java/org/apache/cassandra/cql3/UpdateParameters.java
	src/java/org/apache/cassandra/cql3/statements/AlterTableStatement.java
	src/java/org/apache/cassandra/cql3/statements/AlterTypeStatement.java
	src/java/org/apache/cassandra/cql3/statements/CreateIndexStatement.java
	src/java/org/apache/cassandra/cql3/statements/CreateTableStatement.java
	src/java/org/apache/cassandra/cql3/statements/CreateUserStatement.java
	src/java/org/apache/cassandra/cql3/statements/ModificationStatement.java
	src/java/org/apache/cassandra/cql3/statements/SchemaAlteringStatement.java
	src/java/org/apache/cassandra/cql3/statements/SelectStatement.java
	src/java/org/apache/cassandra/cql3/statements/Selectable.java
	src/java/org/apache/cassandra/db/ArrayBackedSortedColumns.java
	src/java/org/apache/cassandra/db/Cell.java
	src/java/org/apache/cassandra/db/ConsistencyLevel.java
	src/java/org/apache/cassandra/db/DecoratedKey.java
	src/java/org/apache/cassandra/db/DefsTables.java
	src/java/org/apache/cassandra/db/Directories.java
	src/java/org/apache/cassandra/db/ExpiringCell.java
	src/java/org/apache/cassandra/db/OnDiskAtom.java
	src/java/org/apache/cassandra/db/SystemKeyspace.java
	src/java/org/apache/cassandra/db/commitlog/CommitLog.java
	src/java/org/apache/cassandra/db/filter/ColumnCounter.java
	src/java/org/apache/cassandra/db/filter/ColumnSlice.java
	src/java/org/apache/cassandra/db/filter/QueryFilter.java
	src/java/org/apache/cassandra/db/index/SecondaryIndex.java
	src/java/org/apache/cassandra/db/index/SecondaryIndexManager.java
	src/java/org/apache/cassandra/db/marshal/UserType.java
	src/java/org/apache/cassandra/gms/FailureDetector.java
	src/java/org/apache/cassandra/gms/Gossiper.java
	src/java/org/apache/cassandra/io/compress/CompressedSequentialWriter.java
	src/java/org/apache/cassandra/io/compress/CompressionMetadata.java
	src/java/org/apache/cassandra/io/sstable/Descriptor.java
	src/java/org/apache/cassandra/io/sstable/IndexSummaryBuilder.java
	src/java/org/apache/cassandra/io/sstable/SSTableReader.java
	src/java/org/apache/cassandra/io/sstable/SSTableScanner.java
	src/java/org/apache/cassandra/io/sstable/SSTableWriter.java
	src/java/org/apache/cassandra/io/util/AbstractDataOutput.java
	src/java/org/apache/cassandra/io/util/DataOutputPlus.java
	src/java/org/apache/cassandra/io/util/MmappedSegmentedFile.java
	src/java/org/apache/cassandra/io/util/SequentialWriter.java
	src/java/org/apache/cassandra/locator/SimpleSeedProvider.java
	src/java/org/apache/cassandra/net/IAsyncCallbackWithFailure.java
	src/java/org/apache/cassandra/net/OutboundTcpConnection.java
	src/java/org/apache/cassandra/service/MigrationManager.java
	src/java/org/apache/cassandra/service/StorageProxy.java
	src/java/org/apache/cassandra/service/StorageService.java
	src/java/org/apache/cassandra/service/paxos/PaxosState.java
	src/java/org/apache/cassandra/transport/Frame.java
	src/java/org/apache/cassandra/transport/Message.java
	src/java/org/apache/cassandra/transport/Server.java
	src/java/org/apache/cassandra/transport/messages/BatchMessage.java
	src/java/org/apache/cassandra/transport/messages/CredentialsMessage.java
	src/java/org/apache/cassandra/transport/messages/ExecuteMessage.java
	src/java/org/apache/cassandra/transport/messages/OptionsMessage.java
	src/java/org/apache/cassandra/transport/messages/QueryMessage.java
	src/java/org/apache/cassandra/transport/messages/ReadyMessage.java
	src/java/org/apache/cassandra/transport/messages/RegisterMessage.java
	src/java/org/apache/cassandra/transport/messages/ResultMessage.java
	src/java/org/apache/cassandra/transport/messages/StartupMessage.java
	src/java/org/apache/cassandra/transport/messages/SupportedMessage.java
	src/java/org/apache/cassandra/utils/concurrent/SimpleCondition.java

+++ src/java/org/apache/cassandra/cache/AutoSavingCache.java
@@ -36,6 +36,7 @@ import org.apache.cassandra.db.marshal.BytesType;
 import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.Keyspace;
 import org.apache.cassandra.io.FSWriteError;
+import org.apache.cassandra.io.util.DataOutputPlus;
 import org.apache.cassandra.io.util.FileUtils;
 import org.apache.cassandra.io.util.LengthAvailableInputStream;
 import org.apache.cassandra.io.util.SequentialWriter;
@@ -172,6 +173,8 @@ public class AutoSavingCache<K extends CacheKey, V> extends InstrumentingCache<K
                 type = OperationType.KEY_CACHE_SAVE;
             else if (cacheType == CacheService.CacheType.ROW_CACHE)
                 type = OperationType.ROW_CACHE_SAVE;
+            else if (cacheType == CacheService.CacheType.COUNTER_CACHE)
+                type = OperationType.COUNTER_CACHE_SAVE;
             else
                 type = OperationType.UNKNOWN;
 
@@ -258,32 +261,37 @@ public class AutoSavingCache<K extends CacheKey, V> extends InstrumentingCache<K
         {
             File path = getCachePath(pathInfo.keyspace, pathInfo.columnFamily, pathInfo.cfId, CURRENT_VERSION);
             File tmpFile = FileUtils.createTempFile(path.getName(), null, path.getParentFile());
-            return SequentialWriter.open(tmpFile, true);
+            return SequentialWriter.open(tmpFile);
         }
 
         private void deleteOldCacheFiles()
         {
             File savedCachesDir = new File(DatabaseDescriptor.getSavedCachesLocation());
             assert savedCachesDir.exists() && savedCachesDir.isDirectory();
-
-            for (File file : savedCachesDir.listFiles())
+            File[] files = savedCachesDir.listFiles();
+            if (files != null)
             {
-                if (!file.isFile())
-                    continue; // someone's been messing with our directory.  naughty!
-
-                if (file.getName().endsWith(cacheType.toString())
-                    || file.getName().endsWith(String.format("%s-%s.db", cacheType.toString(), CURRENT_VERSION)))
+                for (File file : files)
                 {
-                    if (!file.delete())
-                        logger.warn("Failed to delete {}", file.getAbsolutePath());
+                    if (!file.isFile())
+                        continue; // someone's been messing with our directory.  naughty!
+
+                    if (file.getName().endsWith(cacheType.toString())
+                            || file.getName().endsWith(String.format("%s-%s.db", cacheType.toString(), CURRENT_VERSION)))
+                    {
+                        if (!file.delete())
+                            logger.warn("Failed to delete {}", file.getAbsolutePath());
+                    }
                 }
             }
+            else
+                logger.warn("Could not list files in {}", savedCachesDir);
         }
     }
 
     public interface CacheSerializer<K extends CacheKey, V>
     {
-        void serialize(K key, DataOutput out) throws IOException;
+        void serialize(K key, DataOutputPlus out) throws IOException;
 
         Future<Pair<K, V>> deserialize(DataInputStream in, ColumnFamilyStore cfs) throws IOException;
     }
